Import Data and Peak tables from xslx

import_dataXL() reads sheets from xslx and load them.

Usage

import_dataXL(xlsx_file_name, sheet_name, interactive_output = FALSE)

Arguments

xlsx_file_name

A name of excel file.

sheet_name

A name of indicating the sheet of interest.

interactive_output

Boolean indicating whether to output the table in interactive mode using reactable.

Value

The table of the tibble format or as a reactable if interactive_output is TRUE.

Examples

library(MDAW)
excel_file <- "Gastric_NMR.xlsx"
path <- system.file("extdata", "Gastric_NMR.xlsx", package = "MDAW")
MDAW::import_dataXL(xlsx_file_name = path,
    sheet_name = "data")
#> Loading sheet: data 
#> TOTAL ROWS: 140 
#> Done!
#> # A tibble: 140 × 137
#>    RowID   Idx `Day of Expt`       Sample_Type    QC Batch Order Sample_id    M1
#>    <chr> <dbl> <dttm>              <chr>       <dbl> <dbl> <dbl> <chr>     <dbl>
#>  1 1         1 2014-12-08 00:00:00 QC              1     1     1 sample_1   90.1
#>  2 2         2 2014-12-08 00:00:00 Sample          0     1     2 sample_2   43  
#>  3 3         3 2014-12-08 00:00:00 Sample          0     1     3 sample_3  214. 
#>  4 4         4 2014-12-08 00:00:00 Sample          0     1     4 sample_4   31.6
#>  5 5         5 2014-12-08 00:00:00 Sample          0     1     5 sample_5   81.9
#>  6 6         6 2014-12-08 00:00:00 Sample          0     1     6 sample_6  197. 
#>  7 7         7 2014-12-08 00:00:00 Sample          0     1     7 sample_7   45.5
#>  8 8         8 2014-12-08 00:00:00 Sample          0     1     8 sample_8   91  
#>  9 9         9 2014-12-08 00:00:00 Sample          0     1     9 sample_9  481. 
#> 10 10       10 2014-12-08 00:00:00 QC              1     1    10 sample_10  62.2
#> # ℹ 130 more rows
#> # ℹ 128 more variables: M2 <dbl>, M3 <dbl>, M4 <dbl>, M5 <dbl>, M6 <dbl>,
#> #   M7 <dbl>, M8 <dbl>, M9 <dbl>, M10 <dbl>, M11 <dbl>, M12 <dbl>, M13 <dbl>,
#> #   M14 <dbl>, M15 <dbl>, M16 <dbl>, M17 <dbl>, M18 <dbl>, M19 <dbl>,
#> #   M20 <dbl>, M21 <dbl>, M22 <dbl>, M23 <dbl>, M24 <dbl>, M25 <dbl>,
#> #   M26 <dbl>, M27 <dbl>, M28 <dbl>, M29 <dbl>, M30 <dbl>, M31 <dbl>,
#> #   M32 <dbl>, M33 <dbl>, M34 <dbl>, M35 <dbl>, M36 <dbl>, M37 <dbl>, …