Univariate Statistics for Two-Class Data

Calculates univariate statistics for two-class data including mean, median, confidence intervals, and statistical tests such as t-test or Mann-Whitney U test.

Usage

univariate_2class(
  DataTable,
  PeakTable,
  group,
  posclass,
  parametric = TRUE,
  seed = NULL
)

Arguments

DataTable

Data sheet with the required columns.

PeakTable

Peak sheet with the required columns.

group

Name of the column in the DataTable that contains the Class data.

posclass

Name of the positive class in the group column.

parametric

If TRUE, parametric statistics (mean, t-test) are calculated. If FALSE, non-parametric statistics (median, Mann-Whitney U test) are calculated. i.e. Median 95%CI and Mann-Whitney U test is calculated instead of Mean 95%CI and t-test.

seed

Used to seed the generator for the Median 95% CI bootstrap (resample with replacement) when calculating bootstrap confidence intervals. Ignored if parametric is TRUE. Note, if parametric=TRUE the seed is irrelevant as Mean 95% CI is calculated parametrically.

Value

A data frame containing univariate statistics for each peak. The table should contain multiple univariate statistics (2 class).

Examples

# Example usage
DataTable <- MDAW::table_data
PeakTable <- MDAW::table_peak
table_data_2Class <- DataTable |>
      dplyr::filter(Class == c("HE", "GC"))
# Reduce data table only to GC and HE class members
pos_outcome <- "GC"
univariate_2class(table_data_2Class, PeakTable, group = "Class",
      posclass = pos_outcome)